requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
Django==4.0.6
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.34.4
idna==3.3
jsonpickle==2.2.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
mock==4.0.3
module-name==0.5.1
numpy==1.23.1
numpy-financial==1.0.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scipy==1.9.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.11
xlcalculator==0.4.2
yearfrac==0.4.4
